Day 190: X_{P_n}(q,t) via Hikita recipe for n=2,3 + AP2018 locator correction
- scripts/day190/qt_hikita_P2_P3.py: SymPy computation of X_{P_2}(q,t) = t(1+t) e_2
  and X_{P_3}(q,t) = t^3(1+t+t^2) e_3 + t^2 (e_1 star e_2), matched to Hikita Ex 4.6.
- 2026-09-11-day190-qt-Hikita-P2-P3.md: prose writeup with sanity checks; (Re) does
  NOT lift cleanly to star-product (P_n not disjoint union).
- writing/2026-09-11-day190-ap-locator-correction.tex/.pdf: correction of Day 188
  retraction locator (AP2018 Thm 38 eq 22 does not contain (Re) as claimed;
  Clio Day 184 review + independent fact-check both flagged).

+++ scratch/day184/log_concavity_bk.py
@@ -20,12 +20,12 @@ def main() -> None:
         Integer(417),
         Integer(7851),
         Integer(164124),
-        Integer(3663984),
-        Integer(85498458),
-        Integer(2058089283),
-        Integer(50705502591),
-        Integer(1272084879132),
-        Integer(32365470683334),
+        Integer(3661389),
+        Integer(85384566),
+        Integer(2056373739),
+        Integer(50751637140),
+        Integer(1276862920140),
+        Integer(32626363346505),
     ]
 
     n = len(b)
